move metric/size to thor
