+++ block/block.go
@@ -11,7 +11,7 @@ import (
 	"sync/atomic"
 
 	"github.com/ethereum/go-ethereum/rlp"
-	"github.com/vechain/thor/v2/metric"
+	"github.com/vechain/thor/v2/thor"
 	"github.com/vechain/thor/v2/tx"
 )
 
@@ -90,16 +90,16 @@ func (b *Block) DecodeRLP(s *rlp.Stream) error {
 		header: &payload.Header,
 		txs:    payload.Txs,
 	}
-	b.cache.size.Store(metric.StorageSize(rlp.ListSize(size)))
+	b.cache.size.Store(thor.StorageSize(rlp.ListSize(size)))
 	return nil
 }
 
 // Size returns block size in bytes.
-func (b *Block) Size() metric.StorageSize {
+func (b *Block) Size() thor.StorageSize {
 	if cached := b.cache.size.Load(); cached != nil {
-		return cached.(metric.StorageSize)
+		return cached.(thor.StorageSize)
 	}
-	var size metric.StorageSize
+	var size thor.StorageSize
 	rlp.Encode(&size, b)
 	b.cache.size.Store(size)
 	return size
